* Refatoração de código: atualização dos testes * Tentativa de mostrar os dados do coverage no SonarQube
