+++ app/src/androidTest/java/com/nutricampus/app/acceptance/Animal02BuscarActivityTest.java
@@ -19,6 +19,7 @@ import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 import org.hamcrest.core.IsInstanceOf;
 import org.junit.FixMethodOrder;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -51,7 +52,8 @@ import static org.hamcrest.Matchers.allOf;
 @android.support.test.filters.LargeTest
 @RunWith(AndroidJUnit4.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class AnimalBuscarActivityTest {
+@Ignore
+public class Animal02BuscarActivityTest {
         private Activity currentActivity;
 
     @Rule
