+++ app/src/androidTest/java/com/nutricampus/app/acceptance/Animal04AtivoInativoActivityTest.java
@@ -17,6 +17,7 @@ import com.nutricampus.app.database.SharedPreferencesManager;
 import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
+import org.hamcrest.core.IsInstanceOf;
 import org.junit.FixMethodOrder;
 import org.junit.Rule;
 import org.junit.Test;
@@ -27,11 +28,13 @@ import java.util.Collection;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
+import static android.support.test.espresso.Espresso.pressBack;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.closeSoftKeyboard;
 import static android.support.test.espresso.action.ViewActions.longClick;
 import static android.support.test.espresso.action.ViewActions.replaceText;
 import static android.support.test.espresso.action.ViewActions.scrollTo;
+import static android.support.test.espresso.assertion.ViewAssertions.matches;
 import static android.support.test.espresso.contrib.RecyclerViewActions.actionOnItemAtPosition;
 import static android.support.test.espresso.matcher.ViewMatchers.isDisplayed;
 import static android.support.test.espresso.matcher.ViewMatchers.withContentDescription;
@@ -42,38 +45,59 @@ import static android.support.test.runner.lifecycle.Stage.RESUMED;
 import static org.hamcrest.Matchers.allOf;
 
 /**
- * Created by jorge on 23/07/17.
+ * Created by jorge on 25/07/17.
  */
 
 @java.lang.SuppressWarnings("squid:S2925") //  SonarQube ignora o sleep())
 @android.support.test.filters.LargeTest
 @RunWith(AndroidJUnit4.class)
 @FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class AnimalExcluirActivityTest {
+public class Animal04AtivoInativoActivityTest {
     private Activity currentActivity;
     @Rule
     public ActivityTestRule<LoginActivity> mActivityTestRule = new ActivityTestRule<>(LoginActivity.class);
 
     @Test
-    public void animalCadastroActivityTest2() throws Exception {//Excluir Animal ("Sim")
+    public void animalEditarActivityTest4() throws Exception {//Alterando animal para ativo ou inativo
         prepararTeste();
-        closeKeyboard();
-        ViewInteraction recyclerView2 = onView(
-                allOf(withId(R.id.material_drawer_recycler_view),
-                        withParent(allOf(withId(R.id.material_drawer_slider_layout),
-                                withParent(withId(R.id.material_drawer_layout)))),
-                        isDisplayed()));
-        recyclerView2.perform(actionOnItemAtPosition(5, click()));
 
         onView(withText("Flor")).perform(longClick());
 
         ViewInteraction appCompatTextView6 = onView(
-                allOf(withId(android.R.id.title), withText("Excluir"), isDisplayed()));
+                allOf(withId(android.R.id.title), withText("Editar"), isDisplayed()));
         appCompatTextView6.perform(click());
+        pressBack();
 
-        ViewInteraction appCompatButton22 = onView(
-                allOf(withId(android.R.id.button1), withText("Sim")));
-        appCompatButton22.perform(scrollTo(), click());
+        closeKeyboard();
+
+        Thread.sleep(1500);
+        onView(withId(R.id.switch_ativo)).perform(click());
+        Thread.sleep(4000);
+        onView(withId(R.id.switch_ativo)).perform(click());
+        Thread.sleep(4000);
+
+        ViewInteraction appCompatButton8 = onView(
+                allOf(withId(R.id.btnConfimarDados), withText("Confirmar dados"),
+                        withParent(allOf(withId(R.id.fragmentDadosAnimal),
+                                withParent(withId(R.id.pager)))),
+                        isDisplayed()));
+        appCompatButton8.perform(click());
+
+        ViewInteraction appCompatButton10 = onView(
+                allOf(withId(R.id.btn_salvar)));
+        appCompatButton10.perform(scrollTo(), click());
+        closeKeyboard();
+        ViewInteraction textView4 = onView(
+                allOf(withId(android.R.id.text1), withText("Procure uma propriedade"),
+                        childAtPosition(
+                                allOf(withId(R.id.spinnerPropriedade),
+                                        childAtPosition(
+                                                IsInstanceOf.<View>instanceOf(android.widget.LinearLayout.class),
+                                                0)),
+                                0),
+                        isDisplayed()));
+        textView4.check(matches(withText("Procure uma propriedade")));
+
     }
     public void prepararTeste()throws Exception{
         doLogout();
@@ -94,6 +118,13 @@ public class AnimalExcluirActivityTest {
                         withParent(withId(R.id.toolbar)),
                         isDisplayed()));
         appCompatImageButton.perform(click());
+
+        ViewInteraction recyclerView2 = onView(
+                allOf(withId(R.id.material_drawer_recycler_view),
+                        withParent(allOf(withId(R.id.material_drawer_slider_layout),
+                                withParent(withId(R.id.material_drawer_layout)))),
+                        isDisplayed()));
+        recyclerView2.perform(actionOnItemAtPosition(5, click()));
     }
 
     private static Matcher<View> childAtPosition(
@@ -141,3 +172,8 @@ public class AnimalExcluirActivityTest {
         }
     }
 }
+
+
+
+
+
