+++ app/src/androidTest/java/com/nutricampus/app/acceptance/Animal05HistoricoActivityTest.java
@@ -1,10 +1,12 @@
 package com.nutricampus.app.acceptance;
 
 import android.app.Activity;
+import android.support.design.widget.TextInputLayout;
 import android.support.test.espresso.ViewInteraction;
 import android.support.test.rule.ActivityTestRule;
 import android.support.test.runner.AndroidJUnit4;
 import android.support.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
+import android.test.suitebuilder.annotation.LargeTest;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.ViewParent;
@@ -18,17 +20,15 @@ import org.hamcrest.Description;
 import org.hamcrest.Matcher;
 import org.hamcrest.TypeSafeMatcher;
 import org.hamcrest.core.IsInstanceOf;
-import org.junit.FixMethodOrder;
+import org.junit.Ignore;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.junit.runners.MethodSorters;
 
 import java.util.Collection;
 
 import static android.support.test.InstrumentationRegistry.getInstrumentation;
 import static android.support.test.espresso.Espresso.onView;
-import static android.support.test.espresso.Espresso.pressBack;
 import static android.support.test.espresso.action.ViewActions.click;
 import static android.support.test.espresso.action.ViewActions.closeSoftKeyboard;
 import static android.support.test.espresso.action.ViewActions.longClick;
@@ -49,51 +49,49 @@ import static org.hamcrest.Matchers.allOf;
  */
 
 @java.lang.SuppressWarnings("squid:S2925") //  SonarQube ignora o sleep())
-@android.support.test.filters.LargeTest
+@LargeTest
 @RunWith(AndroidJUnit4.class)
-@FixMethodOrder(MethodSorters.NAME_ASCENDING)
-public class AnimalAtivoInativoActivityTest {
+@Ignore // Animal "Florinda" não cadastrado anteriormente
+public class Animal05HistoricoActivityTest {
     private Activity currentActivity;
     @Rule
     public ActivityTestRule<LoginActivity> mActivityTestRule = new ActivityTestRule<>(LoginActivity.class);
 
     @Test
-    public void animalEditarActivityTest4() throws Exception {//Alterando animal para ativo ou inativo
+    public void HistoricoAnimal1() throws  Exception {
         prepararTeste();
-
-        onView(withText("Flor")).perform(longClick());
-
-        ViewInteraction appCompatTextView6 = onView(
-                allOf(withId(android.R.id.title), withText("Editar"), isDisplayed()));
-        appCompatTextView6.perform(click());
-        pressBack();
-        onView(withId(R.id.switch_ativo)).perform(click());
-        Thread.sleep(4000);
-        onView(withId(R.id.switch_ativo)).perform(click());
-        Thread.sleep(4000);
-        ViewInteraction appCompatButton8 = onView(
-                allOf(withId(R.id.btnConfimarDados), withText("Confirmar dados"),
-                        withParent(allOf(withId(R.id.fragmentDadosAnimal),
-                                withParent(withId(R.id.pager)))),
-                        isDisplayed()));
-        appCompatButton8.perform(click());
-
-        ViewInteraction appCompatButton10 = onView(
-                allOf(withId(R.id.btn_salvar)));
-        appCompatButton10.perform(scrollTo(), click());
+        onView(withText("Florinda")).perform(longClick());
         closeKeyboard();
-        ViewInteraction textView4 = onView(
-                allOf(withId(android.R.id.text1), withText("Procure uma propriedade"),
+        ViewInteraction appCompatTextView6 = onView(
+                allOf(withId(android.R.id.title), withText("Ver histórico de dados"), isDisplayed()));
+        appCompatTextView6.perform(click());
+        closeKeyboard();
+        ViewInteraction relativeLayout = onView(
+                allOf(withId(R.id.ItensListaAnimal),
                         childAtPosition(
-                                allOf(withId(R.id.spinnerPropriedade),
-                                        childAtPosition(
-                                                IsInstanceOf.<View>instanceOf(android.widget.LinearLayout.class),
-                                                0)),
+                                allOf(withId(R.id.listDadosCompl),
+                                        withParent(withId(R.id.telaListaDadosCompl))),
                                 0),
                         isDisplayed()));
-        textView4.check(matches(withText("Procure uma propriedade")));
+        relativeLayout.perform(click());
+
+        ViewInteraction editText2 = onView(
+                allOf(withId(R.id.input_peso_vivo), withText("500.0"),
+                        childAtPosition(
+                                childAtPosition(
+                                        IsInstanceOf.<View>instanceOf(TextInputLayout.class),
+                                        0),
+                                0),
+                        isDisplayed()));
+        editText2.check(matches(withText("500.0")));
+
+        ViewInteraction appCompatButton25 = onView(
+                allOf(withId(R.id.btn_salvar), withText("Corrigir")));
+        appCompatButton25.perform(scrollTo(), click());
+        closeKeyboard();
 
     }
+
     public void prepararTeste()throws Exception{
         doLogout();
         ViewInteraction appCompatEditText = onView(
@@ -167,8 +165,3 @@ public class AnimalAtivoInativoActivityTest {
         }
     }
 }
-
-
-
-
-
