+++ app/src/androidTest/java/com/nutricampus/app/acceptance/ProducaoDeLeiteExclusaoAcceptanceTest.java
@@ -23,6 +23,8 @@ import com.nutricampus.app.entities.ProducaoDeLeite;
 import com.nutricampus.app.entities.Propriedade;
 import com.nutricampus.app.entities.Proprietario;
 
+import org.junit.After;
+import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -68,9 +70,14 @@ public class ProducaoDeLeiteExclusaoAcceptanceTest {
         return currentActivity;
     }
 
-    @Test
-    public void producaoDeLeiteExclusaoTA1() {
-        String admin = "admin";
+
+    private RepositorioProducaoDeLeite repositorioProducaoDeLeite;
+    private ProducaoDeLeite producaoDeLeite1;
+    private ProducaoDeLeite producaoDeLeite2;
+
+    @Before
+    public void preparaDados() {
+
         String nome1 = "nomeAdmin proprietario", nome2 = "nomeAdmin proprietario";
         String email1 = "email@proprietario1.com", email2 = "email@proprietario2.com";
         String cpf1 = "000.000.000-000", cpf2 = "999.999.999-99";
@@ -81,38 +88,67 @@ public class ProducaoDeLeiteExclusaoAcceptanceTest {
         String bairro1 = "bairro1", bairro2 = "bairro2";
         String cep1 = "00000-000", cep2 = "99999-999";
 
+        int idAnimal1;
+        int idAnimal2;
+        int idPropriedade1;
+        int idPropriedade2;
+        int idProprietario1 = 0;
+        int idProprietario2 = 0;
+
         RepositorioProprietario repositorioProprietario = new RepositorioProprietario(InstrumentationRegistry.getTargetContext());
         if (repositorioProprietario.buscarProprietario(cpf1) == null) {
-            repositorioProprietario.inserirProprietario(new Proprietario(1, cpf1, nome1, email1, tel1));
-            repositorioProprietario.inserirProprietario(new Proprietario(2, cpf2, nome2, email2, tel2));
+            idProprietario1 = repositorioProprietario.inserirProprietario(new Proprietario(1, cpf1, nome1, email1, tel1));
+            idProprietario2 = repositorioProprietario.inserirProprietario(new Proprietario(2, cpf2, nome2, email2, tel2));
         }
         RepositorioPropriedade repositorioPropriedade = new RepositorioPropriedade(InstrumentationRegistry.getTargetContext());
         if (repositorioPropriedade.buscarPropriedade(prop1) == null) {
-            repositorioPropriedade.inserirPropriedade(new Propriedade(1, prop1, tel1, rua1, bairro1, cep1, "Garanhuns", "Pernambuco", "000", 1, 1));
-            repositorioPropriedade.inserirPropriedade(new Propriedade(2, prop2, tel2, rua2, bairro2, cep2, "Caruaru", "Pernambuco", "999", 2, 1));
+            idPropriedade1 = repositorioPropriedade.inserirPropriedade(new Propriedade(1, prop1, tel1, rua1, bairro1, cep1, "Garanhuns", "Pernambuco", "000", idProprietario1, 1));
+            idPropriedade2 = repositorioPropriedade.inserirPropriedade(new Propriedade(2, prop2, tel2, rua2, bairro2, cep2, "Caruaru", "Pernambuco", "999", idProprietario2, 1));
+        } else {
+            idPropriedade1 = repositorioPropriedade.buscarPropriedade(prop1).getId();
+            idPropriedade2 = repositorioPropriedade.buscarPropriedade(prop2).getId();
         }
+
+
         RepositorioAnimal repoAnimal = new RepositorioAnimal(InstrumentationRegistry.getTargetContext());
-        if (repoAnimal.buscarAnimal(animal1, 1) == null) {
-            repoAnimal.inserirAnimal(new Animal(1, animal1, 1, Calendar.getInstance(), true));
-            repoAnimal.inserirAnimal(new Animal(2, animal2, 2, Calendar.getInstance(), true));
+
+        if (repoAnimal.buscarAnimal(animal1, idPropriedade1) == null) {
+
+            idAnimal1 = repoAnimal.inserirAnimal(new Animal(1, animal1, idPropriedade1, Calendar.getInstance(), true));
+            idAnimal2 = repoAnimal.inserirAnimal(new Animal(2, animal2, idPropriedade2, Calendar.getInstance(), true));
 
             RepositorioDadosComplAnimal repositorioDadosComplAnimal = new RepositorioDadosComplAnimal(InstrumentationRegistry.getTargetContext());
             repositorioDadosComplAnimal.inserirDadosComplAnimal(new DadosComplAnimal(
-                    Calendar.getInstance(), 1, 100, 150, 50, 60, 5, true, true, true, true
+                    Calendar.getInstance(), idAnimal1, 100, 150, 50, 60, 5, true, true, true, true
             ));
             repositorioDadosComplAnimal.inserirDadosComplAnimal(new DadosComplAnimal(
-                    Calendar.getInstance(), 2, 100, 150, 50, 60, 5, true, true, true, true
+                    Calendar.getInstance(), idAnimal2, 100, 150, 50, 60, 5, true, true, true, true
             ));
+        } else {
+            idAnimal1 = repoAnimal.buscarAnimal(animal1, idPropriedade1).getId();
         }
+
         Calendar calendarJunho = Calendar.getInstance();
         calendarJunho.set(2017, Calendar.JUNE, 1);
         Calendar calendarJulho = Calendar.getInstance();
         calendarJulho.set(2017, Calendar.JULY, 1);
-        RepositorioProducaoDeLeite repositorioProducaoDeLeite = new RepositorioProducaoDeLeite(InstrumentationRegistry.getTargetContext());
-        ProducaoDeLeite producaoDeLeite1 = new ProducaoDeLeite(calendarJunho, 1, 999, 999, 999, 999, 999);
-        ProducaoDeLeite producaoDeLeite2 = new ProducaoDeLeite(calendarJulho, 1, 888, 888, 888, 888, 888);
+        repositorioProducaoDeLeite = new RepositorioProducaoDeLeite(InstrumentationRegistry.getTargetContext());
+        producaoDeLeite1 = new ProducaoDeLeite(calendarJunho, idAnimal1, 999, 999, 999, 999, 999);
+        producaoDeLeite2 = new ProducaoDeLeite(calendarJulho, idAnimal1, 888, 888, 888, 888, 888);
         repositorioProducaoDeLeite.inserirProducaoDeLeite(producaoDeLeite1);
         repositorioProducaoDeLeite.inserirProducaoDeLeite(producaoDeLeite2);
+
+    }
+
+    @After
+    public void deleteDados() {
+
+        repositorioProducaoDeLeite.removerProducaoDeLeite(producaoDeLeite1);
+        repositorioProducaoDeLeite.removerProducaoDeLeite(producaoDeLeite2);
+    }
+
+    @Test
+    public void producaoDeLeiteExclusaoTA1() throws InterruptedException {
 
         if (getActivityInstance() instanceof MainActivity) {
             //dummy if
@@ -181,6 +217,10 @@ public class ProducaoDeLeiteExclusaoAcceptanceTest {
                 allOf(withId(android.R.id.text1), withText("Julho"), isDisplayed()));
         appCompatTextView5.perform(click());
 
+        ViewInteraction appCompatTextView55 = onView(
+                allOf(withId(R.id.btn_filtrar_producao), isDisplayed()));
+        appCompatTextView55.perform(click());
+        Thread.sleep(1000);
 
         onView(withText("Peso 888.00 kg"))
                 .perform(longClick());
